requirements.txt
firebase_admin==5.2.0
python-dotenv